+++ crates/ndc-clickhouse/src/sql/query_builder/error.rs
@@ -8,149 +8,98 @@ use ndc_sdk::{
         ObjectTypeName, RelationshipName,
     },
 };
-use std::fmt;
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, thiserror::Error)]
 pub enum QueryBuilderError {
     /// A relationship referenced in the query is missing from the collection_relationships map
+    #[error("Missing relationship: {0}")]
     MissingRelationship(RelationshipName),
     /// An argument required for a native query was not supplied
+    #[error("Argument {argument} required for native query {query} was not supplied")]
     MissingNativeQueryArgument {
         query: CollectionName,
         argument: ArgumentName,
     },
     /// A table was referenced but not found in configuration
+    #[error("Unable to find table {0} in config")]
     UnknownTable(CollectionName),
     /// An argument was supplied for a table that does not have that argument
+    #[error("Unknown argument {argument} supplied for table {table}")]
     UnknownTableArgument {
         table: CollectionName,
         argument: ArgumentName,
     },
-    /// An argument was supplied for a table that does not have that argument
+    /// An argument was supplied for a query that does not have that argument
+    #[error("Unknown argument {argument} supplied for query {query}")]
     UnknownQueryArgument {
         query: CollectionName,
         argument: ArgumentName,
     },
     /// A table in configuration referenced a table type that could not be found
+    #[error("Unable to find table type {0} in config")]
     UnknownTableType(ObjectTypeName),
     /// A column was referenced but not found in configuration
+    #[error("Unable to find column {0} for table {1} in config")]
     UnknownColumn(FieldName, ObjectTypeName),
     /// A field was referenced but not found in configuration
+    #[error("Unknown field {field_name} in type {data_type}")]
     UnknownSubField {
         field_name: FieldName,
         data_type: ClickHouseDataType,
     },
     /// Unable to serialize variables into a json string
+    #[error("Unable to serialize variables into a json string: {0}")]
     CannotSerializeVariables(String),
     /// An unknown single column aggregate function was referenced
+    #[error("Unknown single column aggregate function: {0}")]
     UnknownSingleColumnAggregateFunction(AggregateFunctionName),
     /// An unknown binary comparison operator was referenced
+    #[error("Unknown binary comparison operator: {0}")]
     UnknownBinaryComparisonOperator(ComparisonOperatorName),
     /// A feature is not supported
+    #[error("Not supported: {0}")]
     NotSupported(String),
     /// An error that should never happen, and indicates a bug if triggered
+    #[error("Unexpected: {0}")]
     Unexpected(String),
     /// There was an issue creating typecasting strings
+    #[error("Typecasting: {0}")]
     Typecasting(TypeStringError),
     /// Column type did not match type asserted by request
+    #[error("Column Type Mismatch: expected {expected}, got {got}")]
     ColumnTypeMismatch { expected: String, got: String },
     /// Attempted to cast a JSON value to a mismatching data type
+    #[error("Cannot cast value `{value}` to type `{data_type}`")]
     UnsupportedParameterCast {
         value: serde_json::Value,
         data_type: ParameterType,
     },
     /// Attempted to cast a value to a tuple with a mismatching length
+    #[error("Tuple `{data_type}` length does not match value {value}")]
     TupleLengthMismatch {
         value: serde_json::Value,
         data_type: ParameterType,
     },
     /// Attempted to cast an Array to a named tuple, which should be represented as an object
+    #[error("Expected anonymous tuple for value `{value}`, got `{data_type}`")]
     ExpectedAnonymousTuple {
         value: serde_json::Value,
         data_type: ParameterType,
     },
     /// Attempted to cast an Object to an anonymous tuple, which should be represented as an array
+    #[error("Expected named tuple for value `{value}`, got `{data_type}`")]
     ExpectedNamedTuple {
         value: serde_json::Value,
         data_type: ParameterType,
     },
     /// could not find field required by named tuple or nested in the source json object
+    #[error("Missing field `{field}` for `{data_type}` in `{value}`")]
     MissingNamedField {
         value: serde_json::Value,
         data_type: ParameterType,
         field: String,
     },
 }
-
-impl fmt::Display for QueryBuilderError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            QueryBuilderError::MissingRelationship(rel) => write!(f, "Missing relationship: {rel}"),
-            QueryBuilderError::MissingNativeQueryArgument { query, argument } => write!(
-                f,
-                "Argument {argument} required for native query {query} was not supplied"
-            ),
-            QueryBuilderError::UnknownTable(t) => write!(f, "Unable to find table {t} in config"),
-            QueryBuilderError::UnknownTableArgument { table, argument } => {
-                write!(f, "Unknown argument {argument} supplied for table {table}")
-            }
-            QueryBuilderError::UnknownQueryArgument { query, argument } => {
-                write!(f, "Unknown argument {argument} supplied for query {query}")
-            }
-
-            QueryBuilderError::UnknownTableType(t) => {
-                write!(f, "Unable to find table type {t} in config")
-            }
-            QueryBuilderError::UnknownColumn(c, t) => {
-                write!(f, "Unable to find column {c} for table {t} in config")
-            }
-            QueryBuilderError::UnknownSubField {
-                field_name,
-                data_type,
-            } => {
-                write!(f, "Unknown field {field_name} in type {data_type}")
-            }
-            QueryBuilderError::CannotSerializeVariables(e) => {
-                write!(f, "Unable to serialize variables into a json string: {e}")
-            }
-            QueryBuilderError::UnknownSingleColumnAggregateFunction(agg) => {
-                write!(f, "Unknown single column aggregate function: {agg}")
-            }
-            QueryBuilderError::UnknownBinaryComparisonOperator(op) => {
-                write!(f, "Unknown binary comparison operator: {op}")
-            }
-            QueryBuilderError::NotSupported(e) => write!(f, "Not supported: {e}"),
-            QueryBuilderError::Unexpected(e) => write!(f, "Unexpected: {e}"),
-            QueryBuilderError::Typecasting(e) => write!(f, "Typecasting: {e}"),
-            QueryBuilderError::ColumnTypeMismatch { expected, got } => {
-                write!(f, "Column Type Mismatch: expected {expected}, got {got}")
-            }
-            QueryBuilderError::UnsupportedParameterCast { value, data_type } => {
-                write!(f, "Cannot cast value `{}` to type `{}`", value, data_type)
-            }
-            QueryBuilderError::TupleLengthMismatch { value, data_type } => {
-                write!(f, "Tuple `{data_type}` length does not match value {value}")
-            }
-            QueryBuilderError::ExpectedAnonymousTuple { value, data_type } => {
-                write!(
-                    f,
-                    "Expected anonymous tuple for value `{value}`, got `{data_type}`"
-                )
-            }
-            QueryBuilderError::ExpectedNamedTuple { value, data_type } => write!(
-                f,
-                "Expected named tuple for value `{value}`, got `{data_type}`"
-            ),
-            QueryBuilderError::MissingNamedField {
-                value,
-                data_type,
-                field,
-            } => write!(f, "Missing field `{field}` for `{data_type}` in `{value}`"),
-        }
-    }
-}
-
-impl std::error::Error for QueryBuilderError {}
 
 impl From<QueryBuilderError> for ErrorResponse {
     fn from(value: QueryBuilderError) -> Self {