+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.4] - 2024-10-05
+
+- Implement PrintSchemaAndCapabilities command. Enables the ddn CLI to perform introspection without needing to start an instance of the connector, which makes introspection faster. Note the DDN CLI does not yet implement this.
+- Bug fix: scalar aggregates may return other scalars, creating dependencies. Fix schema response not including those scalar type dependencies 
+
 ## [1.0.3]
 
 - Update SDK version to enable unauthorized access to health endpoint